requirements.txt
requests
streamlit-lottie
matplotlib
python-dotenv
pymongo
notebook
pandas
openai
azure-core
azure-identity
azure-search-documents
pypdf
docx2txt
faiss-cpu
streamlit
streamlit-chat
langchain
langchain-community
langchain-experimental
langchain-core
langchain_openai
langchain-huggingface
huggingface-hub
google-search-results
langchain_google_community
langchain-text-splitters
streamlit-modal
pgvector
pluggy
agentmemory
pypdf2 
python-docx